holo-ssh-keys: Remove code working around openssh < 6.8
Prior to OpenSSH 6.8 (2015-03-18), ssh-keygen used MD5 for fingerprints.
In 6.8 that became configurable (FingerprintHash, the `-E` flag), and
started using SHA256 by default.

The fingerprints are informational; using a different FingerprintHash has
no effect on the operation of Holo.  However, the tests inspect the full
output; they expect it to use SHA256.

That part of holo-ssh-keys was written in 2015-12; of course Travis wasn't
shipping 6.8 yet.  But I'm pretty sure they are now!